code consolidation
