+++ src/scripts/index.js
@@ -6,10 +6,8 @@ import '../styles/index.scss';
 import './config';
 import 'ckeditor4';
 import registerPlugin from './plugin';
-import registerAdapt from './plugindos';
 
 require('../index.html');
 registerPlugin();
-registerAdapt();
 CKEDITOR.replace('editor');
