+++ src/scripts/registerQueryPluginLibretexts.js
@@ -3,7 +3,6 @@
  */
 import '../styles/index.scss';
 import registerPlugin from './plugin';
-import registerAdapt from './plugindos';
 
 // Adds this plugin to the LibreEditor for later activation
 // this will ensure that `registerPlugin()` be called after
@@ -11,7 +10,5 @@ import registerAdapt from './plugindos';
 LibreEditor.queryPlugin = (config) => {
   registerPlugin();
   config.toolbar[12].push('openLibretextsQueryDialog');
-  registerAdapt();
-  config.toolbar[12].push('openLibretextsAdaptDialog');
 };
